Renames test files
